Move dialogs routes
